# examples/expressjs/api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express E-commerce API",
    "version": "1.0.0",
    "description": "Sample Express.js e-commerce API for Bedrock integration demo"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/v1/users": {
      "get": {
        "summary": "Get all users",
        "description": "Retrieve a list of all users in the system",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of users to return",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of users to skip",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "description": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "User's full name",
                    "minLength": 1
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address"
                  },
                  "age": {
                    "type": "integer",
                    "minimum": 18,
                    "maximum": 120,
                    "description": "User's age"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          }
        }
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "summary": "Get user by ID",
        "description": "Retrieve a specific user by their ID",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "User ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "summary": "Get all products",
        "description": "Retrieve a list of all products",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by product category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_price",
            "in": "query",
            "description": "Minimum price filter",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "name": "max_price",
            "in": "query",
            "description": "Maximum price filter",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "name": "in_stock",
            "in": "query",
            "description": "Filter by stock availability",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders": {
      "post": {
        "summary": "Create a new order",
        "description": "Create a new order for a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["user_id", "items", "shipping_address"],
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "description": "ID of the user placing the order"
                  },
                  "items": {
                    "type": "array",
                    "description": "Items in the order",
                    "items": {
                      "type": "object",
                      "required": ["product_id", "quantity"],
                      "properties": {
                        "product_id": {
                          "type": "integer",
                          "description": "Product ID"
                        },
                        "quantity": {
                          "type": "integer",
                          "minimum": 1,
                          "description": "Quantity of the product"
                        }
                      }
                    }
                  },
                  "shipping_address": {
                    "type": "object",
                    "required": ["street", "city", "state", "zip_code"],
                    "properties": {
                      "street": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string"
                      },
                      "zip_code": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "404": {
            "description": "User or product not found"
          }
        }
      }
    },
    "/internal/admin/users": {
      "get": {
        "summary": "Admin - Get all users with sensitive data",
        "description": "Internal admin endpoint - should be excluded from AI tools",
        "responses": {
          "200": {
            "description": "Admin user data with sensitive information"
          }
        }
      }
    },
    "/internal/health": {
      "get": {
        "summary": "Internal health check",
        "description": "Internal health check - should be excluded from AI tools",
        "responses": {
          "200": {
            "description": "Health status"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": ["name", "email"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User's full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "User creation timestamp"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Product ID"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "price": {
            "type": "number",
            "description": "Product price"
          },
          "category": {
            "type": "string",
            "description": "Product category"
          },
          "in_stock": {
            "type": "boolean",
            "description": "Whether product is in stock"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "integer",
            "description": "Order ID"
          },
          "user_id": {
            "type": "integer",
            "description": "ID of user who placed the order"
          },
          "total_amount": {
            "type": "number",
            "description": "Total order amount"
          },
          "status": {
            "type": "string",
            "description": "Order status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Order creation timestamp"
          }
        }
      }
    }
  },
  "tags": []
}